+++ config/filesystems.php
@@ -2,46 +2,52 @@
 
 return [
 
-    /*
-    |--------------------------------------------------------------------------
-    | Default Filesystem Disk
-    |--------------------------------------------------------------------------
-    |
-    | Here you may specify the default filesystem disk that should be used
-    | by the framework. The "local" disk, as well as a variety of cloud
-    | based disks are available to your application. Just store away!
-    |
-    */
+  /*
+  |--------------------------------------------------------------------------
+  | Default Filesystem Disk
+  |--------------------------------------------------------------------------
+  |
+  | Here you may specify the default filesystem disk that should be used
+  | by the framework. The "local" disk, as well as a variety of cloud
+  | based disks are available to your application. Just store away!
+  |
+  */
 
-    'default' => env('FILESYSTEM_DISK', 'public'),
-    'cloud' => env('FILESYSTEM_CLOUD', 'spaces'),
+    'default'  => env('FILESYSTEM_DISK', 'public'),
+    'cloud'    => env('FILESYSTEM_CLOUD', 'spaces'),
     'endpoint' => env('DO_SPACES_ENDPOINT', 'spaces'),
 
-    /*
-    |--------------------------------------------------------------------------
-    | Filesystem Disks
-    |--------------------------------------------------------------------------
-    |
-    | Here you may configure as many filesystem "disks" as you wish, and you
-    | may even configure multiple disks of the same driver. Defaults have
-    | been set up for each driver as an example of the required values.
-    |
-    | Supported Drivers: "local", "ftp", "sftp", "s3"
-    |
-    */
+  /*
+  |--------------------------------------------------------------------------
+  | Filesystem Disks
+  |--------------------------------------------------------------------------
+  |
+  | Here you may configure as many filesystem "disks" as you wish, and you
+  | may even configure multiple disks of the same driver. Defaults have
+  | been set up for each driver as an example of the required values.
+  |
+  | Supported Drivers: "local", "ftp", "sftp", "s3"
+  |
+  */
 
     'disks' => [
 
         'local' => [
             'driver' => 'local',
-            'root' => storage_path('app'),
-            'throw' => false,
+            'root'   => storage_path('app'),
+            'throw'  => false,
         ],
 
         'local-video-chunks' => [
             'driver' => 'scoped',
-            'disk' => 'public',
+            'disk'   => 'public',
             'prefix' => 'chunks',
+        ],
+
+        'temp-files' => [
+            'driver' => 'scoped',
+            'disk'   => 'local', // Use 'local' as the base disk
+            'prefix' => 'temp-files',
         ],
 
 //        'temp-videos' => [
@@ -59,59 +65,59 @@ return [
         'media' => [
             'driver' => 'spaces',
             'root'   => public_path(env('DO_SPACES_FOLDER')),
-            'url'    => env('DO_SPACES_CDN_ENDPOINT').'/'.env('DO_SPACES_FOLDER'),
+            'url'    => env('DO_SPACES_CDN_ENDPOINT') . '/' . env('DO_SPACES_FOLDER'),
         ],
 
 
         'public' => [
-            'driver' => 'local',
-            'root' => storage_path('app/public'),
-            'url' => env('APP_URL').'/storage',
+            'driver'     => 'local',
+            'root'       => storage_path('app/public'),
+            'url'        => env('APP_URL') . '/storage',
             'visibility' => 'public',
-            'throw' => false,
+            'throw'      => false,
         ],
 
         's3' => [
-            'driver' => 's3',
-            'key' => env('AWS_ACCESS_KEY_ID'),
-            'secret' => env('AWS_SECRET_ACCESS_KEY'),
-            'region' => env('AWS_DEFAULT_REGION'),
-            'bucket' => env('AWS_BUCKET'),
+            'driver'   => 's3',
+            'key'      => env('AWS_ACCESS_KEY_ID'),
+            'secret'   => env('AWS_SECRET_ACCESS_KEY'),
+            'region'   => env('AWS_DEFAULT_REGION'),
+            'bucket'   => env('AWS_BUCKET'),
             'endpoint' => env('AWS_USE_PATH_STYLE_ENDPOINT'),
         ],
 
         'spaces' => [
-         'driver' => 's3',
-         'key' => env('DO_SPACES_KEY'),
-         'secret' => env('DO_SPACES_SECRET'),
-         'endpoint' => env('DO_SPACES_ENDPOINT'),
-         'cdn_endpoint' => env('DO_SPACES_CDN_ENDPOINT'),
-         'use_path_style_endpoint' => env('DO_SPACES_USE_PATH_STYLE_ENDPOINT', false),
-         'url' => env('DO_SPACES_URL'),
-         'folder' => env('DO_SPACES_FOLDER'),
-         'region' => env('DO_SPACES_REGION'),
-         'bucket' => env('DO_SPACES_BUCKET'),
-         'bucket_endpoint' => true,
-         'visibility' => 'public'
+            'driver'                  => 's3',
+            'key'                     => env('DO_SPACES_KEY'),
+            'secret'                  => env('DO_SPACES_SECRET'),
+            'endpoint'                => env('DO_SPACES_ENDPOINT'),
+            'cdn_endpoint'            => env('DO_SPACES_CDN_ENDPOINT'),
+            'use_path_style_endpoint' => env('DO_SPACES_USE_PATH_STYLE_ENDPOINT', false),
+            'url'                     => env('DO_SPACES_URL'),
+            'folder'                  => env('DO_SPACES_FOLDER'),
+            'region'                  => env('DO_SPACES_REGION'),
+            'bucket'                  => env('DO_SPACES_BUCKET'),
+            'bucket_endpoint'         => true,
+            'visibility'              => 'public'
         ],
 
     ],
 
-    /*
-    |--------------------------------------------------------------------------
-    | Symbolic Links
-    |--------------------------------------------------------------------------
-    |
-    | Here you may configure the symbolic links that will be created when the
-    | `storage:link` Artisan command is executed. The array keys should be
-    | the locations of the links and the values should be their targets.
-    |
-    */
+  /*
+  |--------------------------------------------------------------------------
+  | Symbolic Links
+  |--------------------------------------------------------------------------
+  |
+  | Here you may configure the symbolic links that will be created when the
+  | `storage:link` Artisan command is executed. The array keys should be
+  | the locations of the links and the values should be their targets.
+  |
+  */
 
     'links' => [
-        public_path('storage') => storage_path('app/public'),
-        public_path('storage/images') => storage_path('app/public/images'),
-        public_path('storage/videos') => storage_path('app/public/videos'),
+        public_path('storage')                => storage_path('app/public'),
+        public_path('storage/images')         => storage_path('app/public/images'),
+        public_path('storage/videos')         => storage_path('app/public/videos'),
         public_path('storage/profile-photos') => storage_path('app/public/profile-photos'),
     ],